Moved Texture::generate() to TextureUtils.
Cleaning up the Texture class a bit.
