+++ OpenTESArena/src/Interface/ChooseClassCreationPanel.cpp
@@ -24,7 +24,7 @@
 ChooseClassCreationPanel::ChooseClassCreationPanel(Game &game)
 	: Panel(game)
 {
-	this->parchment = Texture::generate(Texture::PatternType::Parchment, 180, 40,
+	this->parchment = TextureUtils::generate(TextureUtils::PatternType::Parchment, 180, 40,
 		game.getTextureManager(), game.getRenderer());
 
 	this->titleTextBox = [&game]()
