+++ OpenTESArena/src/Interface/Texture.h
@@ -3,9 +3,6 @@
 
 // A thin SDL_Texture wrapper.
 
-class Renderer;
-class TextureManager;
-
 struct SDL_Texture;
 
 class Texture
@@ -13,14 +10,6 @@ class Texture
 private:
 	SDL_Texture *texture;
 public:
-	// Generated texture types. These refer to patterns used with pop-ups and buttons.
-	enum class PatternType
-	{
-		Parchment,
-		Dark,
-		Custom1 // Light gray with borders.
-	};
-
 	Texture();
 	Texture(const Texture&) = delete;
 	Texture(Texture &&texture);
@@ -28,10 +17,6 @@ public:
 
 	Texture &operator=(const Texture&) = delete;
 	Texture &operator=(Texture &&texture);
-
-	// Generates a new texture from a pattern.
-	static Texture generate(Texture::PatternType type, int width, int height,
-		TextureManager &textureManager, Renderer &renderer);
 
 	int getWidth() const;
 	int getHeight() const;
